6272642  gcc and ADII don't get along

+++ usr/src/uts/common/io/usb/hcd/ehci/ehci.c
@@ -318,7 +318,7 @@ ehci_attach(dev_info_t		*dip,
 	 * could just go)
 	 */
 	hcdi_args.usba_hcdi_register_iblock_cookie =
-	    (ddi_iblock_cookie_t)(uint64_t)ehcip->ehci_intr_pri;
+	    (ddi_iblock_cookie_t)(uintptr_t)ehcip->ehci_intr_pri;
 
 	if (usba_hcdi_register(&hcdi_args, 0) != DDI_SUCCESS) {
 		(void) ehci_cleanup(ehcip);
